Improvements: A default behaviour was added to the load state feature.
Previously, if the emulator tried to load an nonexisting save state it would cause a panic and display a somewhat descriptive error message.
Now different validations are performed in pro of a better UX. If the emulator can find a certain save state, it just returns a reference to the current state (by doing so, the emulation keeps executing as expected).

Files changes: desktop/lib.rs, chi8_core/chip8.rs