Improvements: Now save states are specific to roms.
Now each time a state is saved it will be stored on a file that has the same name of the ROM used to generate such save state (it will obviously have a different file extension tho'). Such an approach will allow users to have a better UX.

A new parameter was added to the save_state and load_state methods of the CHIP8 struct. Now both functions take the name of the rom as an &str.